added authentication middleware
