+++ api/likes/modal.js
@@ -1,7 +1,8 @@
 const db = require("../../data/db-config")
+const { likes } = require("../../data/tableNames")
 
 const likeNickname = (nickname_id, user_id) => {
-    return db("likes")
+    return db(likes)
         .insert({
             nickname_id,
             user_id
@@ -9,9 +10,9 @@ const likeNickname = (nickname_id, user_id) => {
         .returning("*")
 }
 
-const unlikeNickname = (like_id) => {
-    return db("likes")
-        .where({ like_id })
+const unlikeNickname = (nickname_id, user_id) => {
+    return db(likes)
+        .where({ nickname_id, user_id })
         .del()
         .returning("*")
 }
