+++ api/schemas/schneiderSchema.js
@@ -1,61 +1,1 @@
-const {
-  GraphQLSchema,
-  GraphQLObjectType,
-  GraphQLNonNull,
-  GraphQLInt,
-  GraphQLString,
-} = require("graphql");
-const { NicknameType } = require("./types");
-const nickname = require("../nicknames/modal");
-
-const nicknameSchema = new GraphQLSchema({
-  name: "SchneiderSchema",
-  query: new GraphQLObjectType({
-    name: "schneiderQuery",
-    fields: () => ({
-      nicknames: {
-        type: NicknameType,
-      },
-    }),
-  }),
-  mutation: new GraphQLObjectType({
-    name: "schneiderMutation",
-    fields: () => ({
-      addNickname: {
-        name: "addNickname",
-        type: NicknameType,
-        args: {
-          nickname: { type: new GraphQLNonNull(GraphQLString) },
-          user_id: { type: new GraphQLNonNull(GraphQLString) },
-        },
-        resolve: (parent, args) => {
-          return nickname.addNickname(args);
-        },
-      },
-      updateNickname: {
-        name: "updateNickname",
-        type: NicknameType,
-        args: {
-          nickname_id: { type: GraphQLInt },
-          nickname: { type: GraphQLString },
-          user_id: { type: GraphQLString },
-        },
-        resolve: (parent, args) => {
-          return nickname.updateNickname(args);
-        },
-      },
-      deleteNickname: {
-        name: "deleteNickname",
-        type: NicknameType,
-        args: {
-          nickname_id: { type: GraphQLInt },
-        },
-        resolve: (parent, args) => {
-          return nickname.deleteNickname(args);
-        },
-      },
-    }),
-  }),
-});
-
-module.exports = nicknameSchema;
+99,111,110,115,116,32,123,10,32,32,71,114,97,112,104,81,76,83,99,104,101,109,97,44,10,32,32,71,114,97,112,104,81,76,79,98,106,101,99,116,84,121,112,101,44,10,32,32,71,114,97,112,104,81,76,78,111,110,78,117,108,108,44,10,32,32,71,114,97,112,104,81,76,83,116,114,105,110,103,44,10,32,32,71,114,97,112,104,81,76,73,68,44,10,32,32,71,114,97,112,104,81,76,76,105,115,116,44,10,125,32,61,32,114,101,113,117,105,114,101,40,34,103,114,97,112,104,113,108,34,41,59,10,99,111,110,115,116,32,123,32,78,105,99,107,110,97,109,101,84,121,112,101,44,32,85,115,101,114,84,121,112,101,32,125,32,61,32,114,101,113,117,105,114,101,40,34,46,47,116,121,112,101,115,34,41,59,10,47,47,32,99,111,110,115,116,32,110,105,99,107,110,97,109,101,115,32,61,32,114,101,113,117,105,114,101,40,34,46,46,47,110,105,99,107,110,97,109,101,115,47,109,111,100,97,108,34,41,59,10,10,99,111,110,115,116,32,83,99,104,101,109,97,32,61,32,110,101,119,32,71,114,97,112,104,81,76,83,99,104,101,109,97,40,123,10,32,32,110,97,109,101,58,32,34,83,99,104,110,101,105,100,101,114,83,99,104,101,109,97,34,44,10,32,32,113,117,101,114,121,58,32,110,101,119,32,71,114,97,112,104,81,76,79,98,106,101,99,116,84,121,112,101,40,123,10,32,32,32,32,110,97,109,101,58,32,34,115,99,104,110,101,105,100,101,114,81,117,101,114,121,34,44,10,32,32,32,32,102,105,101,108,100,115,58,32,40,41,32,61,62,32,40,123,10,32,32,32,32,32,32,110,105,99,107,110,97,109,101,115,58,32,123,10,32,32,32,32,32,32,32,32,116,121,112,101,58,32,110,101,119,32,71,114,97,112,104,81,76,76,105,115,116,40,78,105,99,107,110,97,109,101,84,121,112,101,41,44,10,32,32,32,32,32,32,32,32,114,101,115,111,108,118,101,58,32,40,112,97,114,101,110,116,44,32,97,114,103,115,41,32,61,62,32,123,10,32,32,32,32,32,32,32,32,32,32,47,47,32,114,101,116,117,114,110,32,110,105,99,107,110,97,109,101,115,46,103,101,116,85,115,101,114,78,105,99,107,110,97,109,101,115,40,41,10,32,32,32,32,32,32,32,32,125,10,32,32,32,32,32,32,125,44,10,32,32,32,32,125,41,44,10,32,32,125,41,44,10,32,32,109,117,116,97,116,105,111,110,58,32,110,101,119,32,71,114,97,112,104,81,76,79,98,106,101,99,116,84,121,112,101,40,123,10,32,32,32,32,110,97,109,101,58,32,34,115,99,104,110,101,105,100,101,114,77,117,116,97,116,105,111,110,34,44,10,32,32,32,32,102,105,101,108,100,115,58,32,40,41,32,61,62,32,40,123,10,32,32,32,32,32,32,97,100,100,78,105,99,107,110,97,109,101,58,32,123,10,32,32,32,32,32,32,32,32,110,97,109,101,58,32,34,97,100,100,78,105,99,107,110,97,109,101,34,44,10,32,32,32,32,32,32,32,32,116,121,112,101,58,32,78,105,99,107,110,97,109,101,84,121,112,101,44,10,32,32,32,32,32,32,32,32,97,114,103,115,58,32,123,10,32,32,32,32,32,32,32,32,32,32,110,105,99,107,110,97,109,101,58,32,123,32,116,121,112,101,58,32,110,101,119,32,71,114,97,112,104,81,76,78,111,110,78,117,108,108,40,71,114,97,112,104,81,76,83,116,114,105,110,103,41,32,125,44,10,32,32,32,32,32,32,32,32,32,32,117,115,101,114,95,105,100,58,32,123,32,116,121,112,101,58,32,110,101,119,32,71,114,97,112,104,81,76,78,111,110,78,117,108,108,40,71,114,97,112,104,81,76,73,68,41,32,125,44,10,32,32,32,32,32,32,32,32,125,44,10,32,32,32,32,32,32,32,32,114,101,115,111,108,118,101,58,32,40,112,97,114,101,110,116,44,32,97,114,103,115,41,32,61,62,32,123,10,32,32,32,32,32,32,32,32,32,32,114,101,116,117,114,110,32,110,105,99,107,110,97,109,101,46,97,100,100,78,105,99,107,110,97,109,101,40,97,114,103,115,41,59,10,32,32,32,32,32,32,32,32,125,44,10,32,32,32,32,32,32,125,44,10,32,32,32,32,32,32,117,112,100,97,116,101,78,105,99,107,110,97,109,101,58,32,123,10,32,32,32,32,32,32,32,32,110,97,109,101,58,32,34,117,112,100,97,116,101,78,105,99,107,110,97,109,101,34,44,10,32,32,32,32,32,32,32,32,116,121,112,101,58,32,78,105,99,107,110,97,109,101,84,121,112,101,44,10,32,32,32,32,32,32,32,32,97,114,103,115,58,32,123,10,32,32,32,32,32,32,32,32,32,32,110,105,99,107,110,97,109,101,95,105,100,58,32,123,32,116,121,112,101,58,32,71,114,97,112,104,81,76,73,68,32,125,44,10,32,32,32,32,32,32,32,32,32,32,110,105,99,107,110,97,109,101,58,32,123,32,116,121,112,101,58,32,71,114,97,112,104,81,76,83,116,114,105,110,103,32,125,44,10,32,32,32,32,32,32,32,32,32,32,117,115,101,114,95,105,100,58,32,123,32,116,121,112,101,58,32,71,114,97,112,104,81,76,73,68,32,125,44,10,32,32,32,32,32,32,32,32,125,44,10,32,32,32,32,32,32,32,32,114,101,115,111,108,118,101,58,32,40,112,97,114,101,110,116,44,32,97,114,103,115,41,32,61,62,32,123,10,32,32,32,32,32,32,32,32,32,32,114,101,116,117,114,110,32,110,105,99,107,110,97,109,101,46,117,112,100,97,116,101,78,105,99,107,110,97,109,101,40,97,114,103,115,41,59,10,32,32,32,32,32,32,32,32,125,44,10,32,32,32,32,32,32,125,44,10,32,32,32,32,32,32,100,101,108,101,116,101,78,105,99,107,110,97,109,101,58,32,123,10,32,32,32,32,32,32,32,32,110,97,109,101,58,32,34,100,101,108,101,116,101,78,105,99,107,110,97,109,101,34,44,10,32,32,32,32,32,32,32,32,116,121,112,101,58,32,78,105,99,107,110,97,109,101,84,121,112,101,44,10,32,32,32,32,32,32,32,32,97,114,103,115,58,32,123,10,32,32,32,32,32,32,32,32,32,32,110,105,99,107,110,97,109,101,95,105,100,58,32,123,32,116,121,112,101,58,32,71,114,97,112,104,81,76,73,68,32,125,44,10,32,32,32,32,32,32,32,32,125,44,10,32,32,32,32,32,32,32,32,114,101,115,111,108,118,101,58,32,40,112,97,114,101,110,116,44,32,97,114,103,115,41,32,61,62,32,123,10,32,32,32,32,32,32,32,32,32,32,114,101,116,117,114,110,32,110,105,99,107,110,97,109,101,46,100,101,108,101,116,101,78,105,99,107,110,97,109,101,40,97,114,103,115,41,59,10,32,32,32,32,32,32,32,32,125,44,10,32,32,32,32,32,32,125,44,10,32,32,32,32,32,32,97,100,100,85,115,101,114,58,32,123,10,32,32,32,32,32,32,32,32,110,97,109,101,58,32,34,97,100,100,85,115,101,114,34,44,10,32,32,32,32,32,32,32,32,116,121,112,101,58,32,85,115,101,114,84,121,112,101,44,10,32,32,32,32,32,32,32,32,97,114,103,115,58,32,123,10,32,32,32,32,32,32,32,32,32,32,101,109,97,105,108,58,32,123,32,116,121,112,101,58,32,110,101,119,32,71,114,97,112,104,81,76,78,111,110,78,117,108,108,40,71,114,97,112,104,81,76,83,116,114,105,110,103,41,32,125,44,10,32,32,32,32,32,32,32,32,32,32,112,97,115,115,119,111,114,100,58,32,123,32,116,121,112,101,58,32,110,101,119,32,71,114,97,112,104,81,76,78,111,110,78,117,108,108,40,71,114,97,112,104,81,76,83,116,114,105,110,103,41,32,125,44,10,32,32,32,32,32,32,32,32,32,32,117,115,101,114,110,97,109,101,58,32,123,32,116,121,112,101,58,32,110,101,119,32,71,114,97,112,104,81,76,78,111,110,78,117,108,108,40,71,114,97,112,104,81,76,83,116,114,105,110,103,41,32,125,44,10,32,32,32,32,32,32,32,32,125,44,10,32,32,32,32,32,32,32,32,114,101,115,111,108,118,101,58,32,97,115,121,110,99,32,40,112,97,114,101,110,116,44,32,97,114,103,115,41,32,61,62,32,123,10,32,32,32,32,32,32,32,32,32,32,99,111,110,115,116,32,104,97,115,104,32,61,32,98,99,114,121,112,116,46,104,97,115,104,83,121,110,99,40,97,114,103,115,46,112,97,115,115,119,111,114,100,44,32,49,48,41,59,10,32,32,32,32,32,32,32,32,32,32,99,111,110,115,116,32,110,101,119,85,115,101,114,32,61,32,97,119,97,105,116,32,117,115,101,114,46,97,100,100,85,115,101,114,40,123,32,46,46,46,97,114,103,115,44,32,112,97,115,115,119,111,114,100,58,32,104,97,115,104,32,125,41,59,10,32,32,32,32,32,32,32,32,32,32,114,101,116,117,114,110,32,110,101,119,85,115,101,114,59,10,32,32,32,32,32,32,32,32,125,44,10,32,32,32,32,32,32,125,44,10,32,32,32,32,125,41,44,10,32,32,125,41,44,10,125,41,59,10,10,109,111,100,117,108,101,46,101,120,112,111,114,116,115,32,61,32,83,99,104,101,109,97,59,10
