+++ package.json
@@ -23,10 +23,12 @@
   "dependencies": {
     "bcryptjs": "^2.4.3",
     "brcypt": "^1.0.1",
+    "cors": "^2.8.5",
     "dotenv": "^10.0.0",
     "express": "^4.17.1",
     "express-graphql": "^0.12.0",
     "graphql": "^16.0.1",
+    "helmet": "^4.6.0",
     "jsonwebtoken": "^8.5.1",
     "knex": "^0.95.14",
     "sqlite3": "^5.0.2"
